syntax: absorb utf8-ranges crate
This commit brings the utf8-ranges crate into regex-syntax as a utf8
sub-module.

This was done because it was observed that utf8-ranges is effectively
unused outside the context of regex-syntax. It is a very small amount of
code, and fits alongside the rest of regex-syntax. In particular, anyone
building a regex engine using regex-syntax will likely need this code
anyway.